+++ glicko/system.h
@@ -5,13 +5,13 @@
 
 class System {
 
-  float start_sigma2, start_rho, start_nu2, sigma2_const, nu2_const, min_rho,
+  double start_sigma2, start_rho, start_nu2, sigma2_const, nu2_const, min_rho,
       min_sigma2, min_nu2;
 
 public:
-  System(float start_sigma2, float start_rho, float start_nu2,
-         float sigma2_const, float nu2_const, float min_rho,
-         float min_sigma2, float min_nu2);
+  System(double start_sigma2, double start_rho, double start_nu2,
+         double sigma2_const, double nu2_const, double min_rho,
+         double min_sigma2, double min_nu2);
   ~System() = default;
 
   friend class Person;
